+++ tests/vm/tstring_openarray.nim
@@ -13,6 +13,7 @@ proc set_all[T](s: var openArray[T]; val: T) =
     s[i] = val
 
 proc test() =
+  block:
     var a0 = "hello_world"
     var a1 = [1,2,3,4,5,6,7,8,9]
     var a2 = @[1,2,3,4,5,6,7,8,9]
@@ -22,6 +23,18 @@ proc test() =
     doAssert a0 == "iiiiiiiiiii"
     doAssert a1 == [4,4,4,4,4,4,4,4,4]
     doAssert a2 == @[4,4,4,4,4,4,4,4,4]
+  block:
+    var a0 = "hi"
+    var b0 = "foobar"
+    when nimvm:
+      discard # otherwise hits: bug #15952
+    else:
+      a0.add b0.toOpenArray(1,3)
+      doAssert a0 == "hioob"
+  proc fn(c: openArray[char]): string =
+    result.add c
+  doAssert fn("def") == "def"
+  doAssert fn(['d','\0', 'f'])[2] == 'f'
 
 const constval0 = "hello".map(proc(x: char): char = x)
 const constval1 = [1,2,3,4].map(proc(x: int): int = x)
@@ -31,4 +44,4 @@ doAssert([1,2,3,4].map(proc(x: int): int = x) == constval1)
 
 test()
 static:
-    test()
+  test()